Pin FastAPI to avoid OpenAPI docs rendering issues (#246)
* Reorder install order in docker build; remove tiles from extension

titiler depends on an earlier version of FastAPI and doesn't allow the
dockerized server to have the FastAPI version dictated by the
stac-fastapi setup.

* Pin fastapi in stac_fastapi.types; remove from other packages.

This pins fastapi to 0.67.* in stac_fastapi.types, and has all other
packages rely on that dependency to pull in the correct version. This
also adds a missing dependency on stac_fastapi.types in
stac_fastapi.api.

Fixes #242

* Remove tiler tests.

This breaks the dockerized tests as titiler isn't installed in the
docker image anymore. Based on the outcome of #245, these (and ideally
more robust tests) can be re-added, or the tiles extension will be
dropped completely.

* Add changelog entry

* Install types first in CI, as it's the base subpackage

* Remove stac_fastapi.extensions[tiles] from test install.

I do not know why, but installing the tiles extra re-installs
stac-fastapi.types in a way that breaks an import.